chore: refactor editor slots
